+++ backend/canisters/user_index/impl/src/updates/remove_bot.rs
@@ -3,7 +3,7 @@ use crate::{RuntimeState, jobs, mutate_state, read_state};
 use canister_api_macros::{proposal, update};
 use canister_tracing_macros::trace;
 use constants::OPENCHAT_BOT_USER_ID;
-use local_user_index_canister::{BotRemoved, BotUninstall, UserIndexEvent};
+use local_user_index_canister::{BotRemoved, UserIndexEvent};
 use types::UserId;
 use user_index_canister::remove_bot::{Response::*, *};
 
@@ -57,13 +57,10 @@ fn remove_bot_impl(args: Args, deleted_by: Option<UserId>, state: &mut RuntimeSt
     );
 
     for (location, details) in bot.installations.iter() {
-        state.data.user_index_event_sync_queue.push(
-            details.local_user_index,
-            UserIndexEvent::BotUninstall(BotUninstall {
-                location: *location,
-                bot_id: args.bot_id,
-            }),
-        );
+        state
+            .data
+            .user_index_event_sync_queue
+            .push(details.local_user_index, UserIndexEvent::BotUninstall(*location, args.bot_id));
     }
 
     jobs::sync_events_to_local_user_index_canisters::try_run_now(state);